requirements/base.txt
discord.py==2.3.*
python-dotenv==1.0.1